requirements.txt
# Personal AI OS - Python Dependencies
# Version: 1.0
#
# 核心依赖
pypdf>=4.0.0
python-docx>=1.1.0

